refactor: move statistics logic to service
- Moved statistics logic to StatisticsService
- Implemented caching for statistics data
- Added cache warming functionality
